aliyunwei/requirements.txt
scikit-learn==0.24.2
pandas==1.3.4
gensim==4.1.2
torchvision==0.12.0
torchtext==0.12.0